Fixed Small issues
Fixed a few small issues
