+++ CharacterCreator/CharacterCreator/Models/Characters.cs
@@ -5,38 +5,24 @@ using System.Threading.Tasks;
 
 namespace CharacterCreator.Models
 {
-    public enum Sex
-    {
-        Male,
-        Female
-    }
 
-    public enum Alignment
-    {
-        LawfulGood,
-        NeutralGood,
-        ChaoticGood,
-        LawfulNeutral,
-        TrueNeutral,
-        ChaoticNeutral,
-        LawfulEvil,
-        NeutralEvil,
-        ChaoticEvil
-    }
     public class Characters
     {
-        public int Id { get; set; }
+        public int CharacterId { get; set; }
         public string Name { get; set; }
         public string Sex { get; set; }
         public string Alignment { get; set; }
         public int Strength { get; set; }
         public int Dexterity { get; set; }
         public int Constitution { get; set; }
-        public int Inteligence { get; set; }
+        public int Intelligence { get; set; }
         public int Wisdom { get; set; }
         public int Charisma { get; set; }
 
         public Class Class { get; set; }
+        public int ClassId { get; set; }
+
         public Race Race { get; set; }
+        public int RaceId { get; set; }
     }
 }
