+++ CharacterCreator/CharacterCreator/Models/Enums/Sex.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace CharacterCreator.Models.Enums
+{
+    public enum Sex
+    {
+        Male,
+        Female
+    }
+}
